Refactor to better set up GPS quality metric calculation
Adds automatic reference station picking
